fix: 마커 탭 시 바텀시트와 충돌 문제 해결

+++ Projects/Features/Map/MapFeatureInterface/Sources/MapView/View/MapViewRepresentable.swift
@@ -61,22 +61,29 @@ struct MapViewRepresentable: UIViewRepresentable {
   }
   
   func updateUIView(_ uiView: NMFNaverMapView, context: Context) {
-    
+    print(context.coordinator.selectedPin?.address)
     if let userLocation = userLocation {
+      print("1")
       moveUserLocation(uiView, to: userLocation, context: context)
     }
     if context.coordinator.markers.isEmpty, !flowerPositions.isEmpty {
+      print("2")
       presentMarkers(uiView, flowers: flowerPositions, context: context)
     }
     // 마커 탭 이벤트 시
     if let data = context.coordinator.selectedPin {
-      drawPathLine(uiView, data: data, for: newPath, context: context)
-    } else {
-      context.coordinator.deletePathMarkers()
+      if newPath != context.coordinator.drawPathPoints {
+        print("3")
+        drawPathLine(uiView, data: data, for: newPath, context: context)
+      } else if context.coordinator.isNeedDeleteMarkers {
+        print("3-1")
+        context.coordinator.deletePathMarkers()
+      }
     }
     
     // 현 위치 재검색 액션
     if requestBounds, !context.coordinator.isInitialBounds {
+      print("4")
       currentVisibleBounds(on: uiView.mapView)
       deleteDrawMarker(context: context)
       requestBounds = false
@@ -84,11 +91,12 @@ struct MapViewRepresentable: UIViewRepresentable {
     
     // 특정 위치에 나타날 데이터가 있을 경우
     if let focusData = focusData {
+      print("5")
       drawPathLine(uiView, data: focusData, for: focusData.path, context: context)
       drawFocusMarker(uiView, result: focusData, context: context)
-      
     } else if context.coordinator.focusData != .none {
       // coordinator에 값이 남아있다면 모두 지움
+      print("5-1")
       context.coordinator.deleteSearchResult()
     }
   }
@@ -166,6 +174,7 @@ extension MapViewRepresentable {
     if marker == context.coordinator.activeMarker { return }
     context.coordinator.deletePathMarkers()
     
+    context.coordinator.isNeedDeleteMarkers = false
     context.coordinator.selectedPin = data
     context.coordinator.activeMarker = marker
     if let onMarkerTapped = onMarkerTapped {
@@ -202,13 +211,14 @@ extension MapViewRepresentable {
   
   /// 경로 선을 그리기 위한 메서드
   private func drawPathLine(_ view: NMFNaverMapView, data: FlowerSpot, for newPath: [MapPoint], context: Context) {
+    
     guard !newPath.isEmpty else {
       context.coordinator.deletePathMarkers()
       return
     }
     // 이미 그려진 경로가 있다면 지우고 다시 그리기
     if context.coordinator.paths != nil { context.coordinator.deletePathMarkers() }
-    
+    context.coordinator.drawPathPoints = newPath
     let lines = newPath.map {
       NMGLatLng(lat: $0.latitude, lng: $0.longitude)
     }
@@ -264,6 +274,7 @@ extension MapViewRepresentable {
       // 마커 탭 이벤트 헨들러
       marker.touchHandler = { (overlay: NMFOverlay) -> Bool in
         if let marker = overlay as? NMFMarker {
+          print("마커 탭")
           marker.iconImage = pin.value.bloomingStatus.activeImage
           markerTapEvent(to: marker, data: pin.value, context: context)
           moveCamera(view, to: position)
